refactor : 패키지 분리 - minor issue
  > CommandUtil 클래스
    - utils 패키지로 이동
