+++ src/main/java/net/woody/utils/CommandUtil.java
@@ -1,4 +1,4 @@
-package net.woody.chess;
+package net.woody.utils;
 
 public class CommandUtil {
